requirements.txt
streamlit 
flask 
pandas 
gspread 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
requests
json 
time